# api-getaway/api/docs/swagger.yaml
definitions:
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  orders.CreateDishRequest:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  orders.Dish:
    properties:
      available:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  orders.GetDishesResponse:
    properties:
      dishes:
        items:
          $ref: '#/definitions/orders.Dish'
        type: array
    type: object
...
  orders.UpdateDishRequest:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
    type: object
...
  users.CreateKitchenRequest:
    properties:
      address:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  users.GetKitchensResponse:
    properties:
      kitchens:
        items:
          $ref: '#/definitions/users.Kitchen'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  users.Kitchen:
    properties:
      address:
        type: string
      created_at:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      phone_number:
        type: string
      rating:
        type: number
      total_orders:
        type: integer
      updated_at:
        type: string
    type: object
  users.SearchKitchensResponse:
    properties:
      kitchens:
        items:
          $ref: '#/definitions/users.Kitchen'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  users.UpdateKitchenRequest:
    properties:
      address:
        type: string
      cuisine_type:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
host: localhost:4444
info:
  contact: {}
  description: This is a sample server for Auth Service.
  title: Auth Service API
  version: "1.0"
paths:
  /api/v1/dishes:
    post:
      consumes:
      - application/json
      description: Create Dishes
      parameters:
      - description: Create Dishes
        in: body
        name: dishes
        required: true
        schema:
          $ref: '#/definitions/orders.CreateDishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Dish'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Dishes
      tags:
      - Dishes
  /api/v1/dishes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Dish
      parameters:
      - description: Delete Dish
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Dish'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete Dish
      tags:
      - Dishes
    get:
      consumes:
      - application/json
      description: Get dish by id
      parameters:
      - description: Get dish by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Dish'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get dish by id
      tags:
      - Dishes
  /api/v1/dishes/kitchen/{id}:
    get:
      consumes:
      - application/json
      description: Get Kitchen Dishes
      parameters:
      - description: Get Kitchen Dishes
        in: path
        name: id
        required: true
        type: string
      - description: Get Kitchen Dishes
        in: query
        name: limit
        type: string
      - description: Get Kitchen Dishes
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.GetDishesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Kitchen Dishes
      tags:
      - Dishes
  /api/v1/dishes/update:
    put:
      consumes:
      - application/json
      description: Update Dish
      parameters:
      - description: Update Dish
        in: body
        name: dish
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateDishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Dish'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update Dish
      tags:
      - Dishes
...
  /api/v1/kitchens:
    get:
      consumes:
      - application/json
      description: List Kitchens
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.GetKitchensResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: List Kitchens
      tags:
      - Kitchens
    post:
      consumes:
      - application/json
      description: Create Kitchen
      parameters:
      - description: Create Kitchen
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/users.CreateKitchenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Kitchen'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Kitchen
      tags:
      - Kitchens
  /api/v1/kitchens/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Kitchen
      parameters:
      - description: Delete Kitchen
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Kitchen'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete Kitchen
      tags:
      - Kitchens
    get:
      consumes:
      - application/json
      description: Get Kitchen
      parameters:
      - description: Get Kitchen by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Kitchen'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Kitchen by id
      tags:
      - Kitchens
    put:
      consumes:
      - application/json
      description: Update Kitchen
      parameters:
      - description: Update Kitchen
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/users.UpdateKitchenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Kitchen'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update Kitchen
      tags:
      - Kitchens
  /api/v1/kitchens/search:
    get:
      consumes:
      - application/json
      description: Search Kitchens
      parameters:
      - description: Search Kitchens
        in: query
        name: name
        type: string
      - description: Cuisine Type
        in: query
        name: cuisine_type
        type: string
      - description: Rating
        in: query
        name: rating
        type: number
      - description: Offset number
        in: query
        name: offset
        type: integer
      - description: Limit per offset
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.SearchKitchensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Search Kitchens
      tags:
      - Kitchens
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"